requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0

# Agent Architecture Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-google-genai>=0.0.1
google-generativeai>=0.0.1

# Additional AI/ML Dependencies
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0

# PDF Processing Dependencies
pypdf2>=3.0.0
unstructured>=0.10.0
pdf2image>=1.16.0
pillow>=10.0.0

# Code Quality Tools
black>=23.0.0
flake8>=6.0.0

